+++ release-notes.md
@@ -5,6 +5,7 @@
 - Add support for ext-transient-seat-v1.
 - Add support for wp-drm-lease-v1.
 - Focus-follows-mouse can now be disabled.
+- Add support for pointer-gestures-unstable-v1.
 
 # 1.1.0 (2024-04-22)